CsvAppendSink and CsvUpsertSink removed, SinkBuilder now takes UpsertPipe and returns Sinks
